refactor(jsi): share the Hermes contract header and the GSD lookup side
Hermes is structurally unlike the other three engines: it exposes the real
facebook::jsi API instead of reimplementing the nativescript::engine shape, so
there is no engine layer to extract. Its public header was already neutral.
What moves is therefore small.

  jsi/hermes/NativeApiJsi.h            - the install/create declaration. Android's
                                         Hermes also exposes facebook::jsi, so this
                                         is shareable verbatim.
  jsi/shared/PreparedSignatureDispatch.h - the GSD lookup side.

The second one completes the split started in the first commit of this branch.
Given a dispatch id, this header finds the generated trampoline; it needs only
the hashing/lookup core, never the metadata walk that computes the ids. So the
line falls in the same place as before: mechanism shared, metadata traversal
per-platform. Its include changed from SignatureDispatchCore.h to
SignatureHashing.h, which is what it actually used.

Forwarding headers at both old paths, so no Apple include changed. The
Apple-side PreparedSignatureDispatch.h forwarder still pulls in
SignatureDispatchCore.h first, because its callers have always had that
transitively.

Verified: iOS TestRunner on QuickJS -- which includes the moved GSD header --
713 specs, 0 failures. The Hermes target compiles, links, and launches.

Its suite cannot be run: Apple Hermes hangs after ~10 specs. That is
pre-existing and unrelated to this branch -- a build at 15e5418c, the commit
before the Android-updates work began, hangs identically. Worth fixing, but as
its own piece of work.

(cherry picked from commit 66a26288a8490c59cd31b9d4afa6390b54810f64)
